requirements.txt
gradio
openai>=1.0.0
huggingface_hub
python-dotenv
